NTP now ignores extra NTP packages sent by the server.
Before this patch, extra NTP packages were queued and caused the system clock
to be set to a time in the past. Now in function tryServer in
rtems_bsdnet_ntp.c old packages are read and discarded before a new NTP
request is sent.

This patch was necessary since it seems that our NTP server sometimes produces
these extra NTP packages.