Bugfix in TFTP driver.
The old implementation reads only enough TFTP packages to satisfy the
application's read request. When read() returns, the TFTP server sends
the next package. Since it doesn't get an acknowledge, it sends the
package again and again. If the application doesn't call read() soon
enough, the server detects a timeout and closes the connection.
In order to avoid this, the file is now read completely and stored in a
filebuffer structure. This structure is allocated on the heap and
released when close() is called on the file.